requirements.txt
cryptography>=41.0.0
Pillow>=10.0.0
numpy>=1.24.0
